0.9 asset extractor ,.,.

+++ assetextractor/AssetExtractor/WikiFormat.cs
@@ -1,3 +1,5 @@
+using UnityEngine.Networking;
+
 namespace AssetExtractor;
 using System;
 using System.Collections.Generic;
@@ -43,12 +45,80 @@ public partial class WikiFormat
 
          public static void exportTexture(Texture texture, String path)
          {
-            File.WriteAllBytes(path, makeReadable(texture).EncodeToPNG());
+             if (File.Exists(path) && AssetExtractor.Instance.labelCopies.Value)
+             {
+                 Log.Debug("file exitst" + (path));
+                    
+                 string pathnew = path.Replace(".png", "");
+                 while (File.Exists(pathnew + ".png"))
+                 {
+                     pathnew += " (copy)";
+                 }
+                 Log.Debug("writing file " + (pathnew));
+                 File.WriteAllBytes(pathnew + ".png", makeReadable(texture).EncodeToPNG());
+             }
+             else
+             {
+                 File.WriteAllBytes(path, makeReadable(texture).EncodeToPNG());
+             }
         }
         
         public static void exportTexture(Sprite sprite, String path)
         {
-            File.WriteAllBytes(path, makeReadable(sprite.texture).EncodeToPNG());
+            Texture2D readableText = makeReadable(sprite.texture);
+            
+            Texture2D newTexture = new Texture2D(Mathf.FloorToInt(sprite.textureRect.width), Mathf.FloorToInt(sprite.textureRect.height));
+            
+            var pixels = readableText.GetPixels(  
+                Mathf.FloorToInt(sprite.textureRect.x), 
+                Mathf.FloorToInt(sprite.textureRect.y), 
+                Mathf.FloorToInt(sprite.textureRect.width), 
+                Mathf.FloorToInt(sprite.textureRect.height) );
+
+            try
+            {
+                newTexture.SetPixels(pixels);
+                newTexture.Apply();
+                
+                if (File.Exists(path) && AssetExtractor.Instance.labelCopies.Value)
+                {
+                    Log.Debug("file exitst" + (path));
+                    
+                    string pathnew = path.Replace(".png", "");
+                    while (File.Exists(pathnew + ".png"))
+                    {
+                        pathnew += " (copy)";
+                    }
+                    Log.Debug("writing file " + (pathnew));
+                    File.WriteAllBytes(pathnew + ".png", newTexture.EncodeToPNG());
+                }
+                else
+                {
+                    File.WriteAllBytes(path, newTexture.EncodeToPNG());
+                }
+            }
+            catch (System.Exception e)
+            {
+                Log.Error(e);
+                Log.Info("exporting png without cropping !!,.., ");
+                if (File.Exists(path) && AssetExtractor.Instance.labelCopies.Value)
+                {
+                    Log.Debug("file exitst" + (path));
+                    
+                    string pathnew = path.Replace(".png", "");
+                    while (File.Exists(pathnew + ".png"))
+                    {
+                        pathnew += " (copy)";
+                    }
+                    Log.Debug("writing file " + (pathnew));
+                    File.WriteAllBytes(pathnew + ".png", readableText.EncodeToPNG());
+                }
+                else
+                {
+                    File.WriteAllBytes(path, readableText.EncodeToPNG());
+                }
+            }
+            
         }
         
         static Texture2D makeReadable(Texture texture)
